Various fixes to grabNewText.php
OMG how did this script even worked?

Important change: Some parts of the script were assuming blindly that local
and remote ids (like page or revision) were the same, while others weren't.
That's not reliable and people using this script should use mwdumper instead
of MediaWiki's importDump.php, which would preserve ids and is more efficient.

PROBLEM: logevents api parameter letype doesn't accept multiple values.
Multiple values work only on wikia but breaks compatibility

SOLVED: by removing the filter, since it doesn't matter to have additional
entries anyway

PROBLEM: On 1.25 inserts on page fail because page_counter was removed from
core in 1.25

SOLVED: Check if wiki supports page counters before attempting to use
this field

PROBLEM: rev_parent_id was always 0 on inserted revisions, resulting in
wrong delta sizes in page histories, and contributions/recent changes
treating each change as page creation.

SOLVED: Now it won't insert a 0 but a null, so you can run
populateParentId.php to fix only the newly added entries.

PROBLEM: minor edits weren't marked as such

SOLVED: Caused by a typo in the $revision variable name when retrieving
the property from api

PROBLEM: If a page is created and then moved within the scope of recent
changes that grabs the script, the move fails because it attempts to do
the move on a not-yet-existing page. Edits are inserted but on the original
page title rather than the new one

SOLVED:
- Ignore page move if the title before the move doesn't exist in our
  database.
- Added processing of log entries in recentchanges to also grab the
  original page if the move left a redirect.
- Page title is now retrieved from api using the page id, to insert
  revisions on the correct title.

PROBLEM: calls to api revisions fails when using continuation parameters
if we provided a start date, since continuation parameters can't be
provided with also a start date

SOLVED: removed the start date after doing the initial call, so next calls
don't have it. Also, provided a smarter way of adding continuation
parameters without checking for a specific parameter name

PROBLEM: rev_sha1 wasn't being populated.
SOLVED: Populate it using Revision::base36Sha1( $text )

PROBLEM: restoring pages didn't work because it was trying to get revisions
on remote wiki based on page id, which we don't have when we're restoring
and the page doesn't exist on the local database
SOLVED: allow processPage to use namespace and title if page id is null

Other cleanups on processPage:
- $result['query']['pages'] is never null even if page does not exist,
  so removing pointless check. Use 'missing' instead
- Now getting page info from api to get additional fields that were
  mostly guessing like lenght, redirect, etc
- page_random was wrong: it should be a number between 0 and 1, but
  we were using rand()...

The content inserted on the text table is now compressed if
$wgCompressRevisions is enabled. The same text id will be reused on the
revision table if the same text exists as an old revision of the same page,
by comparing the rev_sha1 field.

Page restores now inserts deleted revisions from the remote wiki to guard
against selective restoration of pages, so deleted revisions still in the
archive table are kept. Fails gracefully if the user doesn't have permissions
to see deleted revisions.

Now it processes page protections. There was stub code for doing it but
wasn't being in use anywhere.

Moved various globals into member variables.

Using dbw->selectRow instead of select + fetchObject where possible.

Second parameter of $this->error should be a number. Changed from true to 1.

Error if the script fails to login. The script already supports not providing
username and password.

pagesProcessed now contains a list of page ids and not titles, because we use
the page id everywhere, and page moves carry the same page id, so it's safer
to rely on page id than title. Also, since the id changes on page deletions and
restores, we can add it in such cases to not process those pages again.

Using a single DB_MASTER connection instead of a reader and a writer, because
we're doing massive changes to the database and we don't want in any case to
fetch stale data from a lagged slave that may cause an error, or be locked
on an opened transaction by the writer.

Check if user can see deleted revisions for page restores. Otherwise just
ignore them.

updateDeleted is now archiveAndDeletePage and requires the page ID, so we don't
need to query it again on places where we already have it.

Rewrite completely the logic of page moves. In case a page is moved several
times, optionally with deletions and restores involved, or several pages in a
round, the current method fails when the target page already exists and it's
a different one. That's because we use page IDs everywhere, but the logevents
api only returns titles, with the page ID being the current page at that title
that may have been moved after that.

I've created a new method, resolveConflictingTitle, used when
we have to move a page but the target title already exists on the database,
that moves pages recursively, leaving the target title empty for the move.

With this method in mind, the titles involved in a move are processed in the
same manner, comparing page ID and title on the remote wiki to ensure they're
moved to the correct title in one step. This will leave pages in an
inconsistent state if page moves are involved. I'll remove this option because
there's no benefit in having it.

This also touches some other paths like deletions and restores, because it
attempts to process everything in one step. For example, we can't simply delete
a page on a deletion log because it may be the target of a later page move,
where the script has already populated that page. To solve this, check if the
title has been involved in page moves, and just update the deleted revision if
that's the case (only if user has permissions).

Changes tested on a local 1.25 against a remote 1.19, performing:
- page edits
- page moves (with and without redirect), of pages created during the time span
  of the grab, or pages before the last run of the grab
- History merge by moving pages over the same title and restoring them
- History split by deleting, restoring some revisions and moving the page

For testing:
- set up 2 wikis in parallel, performing a backup of their initial state.
- database is backed up every few edits
- restoring each backup, and executing grabNewText.php sequentially
- after each set of restore and grab, the same select on some tables is done on
  both databases, saving the result on a text file and doing a diff of both to
  ensure there are no differencies:
   - revision table: rev_id,rev_page,rev_comment,rev_user,rev_user_text,
     rev_timestamp,rev_minor_edit,rev_deleted,rev_len,rev_parent_id,rev_sha1
   - page table: page_id,page_namespace,page_title,page_restrictions,
     page_is_redirect,page_is_new,page_latest,page_len
- Script also tested without logging in

Change-Id: I14aef39fe896e8d07fa5eba3b2ce546ba9deaf01