GrabNewText: Add option to filter namespaces
This way one can grab changes from a specific list of namespaces.
Useful if you also grabText for specific namespaces and want to
have the content updated for a while until you grab content from
the remaining namespaces

In the processRecentLogs I don't filter on the query itself so I
can check namespace on the target page of the move, in case of
cross-namespace moves.

Change-Id: I46630ab9fd2531ac24184a1f5e6facebd5654c16