MCR, actor table and comments table support. Add inheritance
Support the new slots, content, actor and comments tables.

Class inheritance is great: it reduces the amount of repeated code by placing
common methods in parent classes. Class inheritance exists since a very long
time in PHP, let's use it ;-)

Put the common text things on a single file instead of repeating code
everywhere.

For text grabbers:
Use the new RevisionStore class that does all the dirty stuff in the handful of
tables that we need to touch now. This, of course, may break unexpectedly if
core changes the way we can access those interfaces, but that is now happening
anyway when we fiddle with the database directly...

Sadly, there's no way to do that for the archive table. Or at least I wasn't
able to find one.

All scripts have been tested with local MediaWiki 1.33 against a remote wiki
on 1.32. I haven't tested everything on remote 1.19 but that should still work,
since I haven't touched the way we query info from the api (except some scripts
where the api was rendered incompatible).

bug: T222188
bug: T188710
Change-Id: Ib2e1499f588d7a919a3e7d0eb0f5543b6cd4acb1