Updates after Wikia updating to 1.33
All api pagination (continue parameters) were broken when the target wiki is
more recent than the old 1.19. There was no error... it just failed to detect
that tere were more pages to get from the api after the initial batch.

While I'm on it, I've cheched some of the special-casing --wikia parameters
and removed/updated them to the current era.

Change-Id: I4939ad9171c4a2c190d4d53a52edcc5cea1db7cb